# spec/openapi.yaml
openapi: 3.1.0
info:
  title: x402 Mesh Inference Gateway
  version: 0.1.0
  summary: Pay-per-prompt OpenAI-compatible inference, settled in USDC on Algorand via x402.
  description: |
    A decentralized compute marketplace. Agents call an OpenAI-compatible
    `POST /v1/chat/completions` endpoint; the gateway answers `402 Payment Required` with an
    x402 v2 challenge, the caller re-sends the request carrying an `X-PAYMENT` header, and the
    gateway routes the prompt to the best available community GPU node.

    ## Payment flow

    1. `POST /v1/chat/completions` **without** `X-PAYMENT` → `402` with a `PaymentRequired`
       body. Its `accepts[]` carries the `exact` scheme on Algorand, the USDC ASA id, and the
       price in **atomic units** (`"2000"` = $0.0020 USDC, 6 decimals).
    2. The client builds an atomic Algorand transaction group (the facilitator's fee-payer
       transaction plus the client's ASA transfer), base64-encodes the resulting
       `PaymentPayload` as JSON, and sends it in the `X-PAYMENT` request header.
    3. The gateway verifies and settles through the facilitator, routes the prompt to a node,
       and returns the completion together with a `PAYMENT-RESPONSE` header carrying the
       base64-encoded `SettleResponse`.

    Payment is **gasless for the client**: the facilitator sponsors the ALGO transaction fee
    (advertised as `accepts[].extra.feePayer`), so the caller only signs the USDC transfer leg.

    ## Money representation

...

...

...
servers:
...
security: []
paths:
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      tags: [inference]
      summary: Run a paid chat completion on the mesh.
      description: |
        OpenAI-compatible chat completion. Requires a valid `X-PAYMENT` header; without one the
        gateway responds `402` with the payment challenge describing exactly what to pay.

        Set `stream: true` to receive the completion as Server-Sent Events. The payment is
        verified and settled **before** the first SSE frame is written, so a `200` with
        `Content-Type: text/event-stream` implies the payment already succeeded.

        Unknown request properties are ignored rather than rejected, so OpenAI SDK clients that
        send fields this gateway does not implement (`n`, `presence_penalty`, `user`, …) work
        unchanged. Only validated fields are forwarded to the node.
      security:
        - x402Payment: []
      parameters:
        - $ref: "#/components/parameters/XPaymentHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            examples:
              nonStreaming:
                summary: Buffered completion
                value:
                  model: llama3.1:8b
                  messages:
                    - role: system
                      content: You are a terse assistant.
                    - role: user
                      content: Explain the x402 protocol in one sentence.
                  max_tokens: 256
                  temperature: 0.2
              streaming:
                summary: Server-Sent Events completion
                value:
                  model: llama3.1:8b
                  messages:
                    - role: user
                      content: Stream me a haiku about micropayments.
                  stream: true
      responses:
        "200":
          description: |
            Completion produced and payment settled. `PAYMENT-RESPONSE` carries the base64 JSON
            `SettleResponse` for the inbound (client to gateway) leg.
          headers:
            PAYMENT-RESPONSE:
              $ref: "#/components/headers/PaymentResponse"
            X-Mesh-Node-Id:
              $ref: "#/components/headers/MeshNodeId"
            X-Mesh-Request-Id:
              $ref: "#/components/headers/MeshRequestId"
            X-Mesh-Route-Reason:
              $ref: "#/components/headers/MeshRouteReason"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
              example:
                id: chatcmpl-0f3a9c1f5b2e4d7a
                object: chat.completion
                created: 1785772800
                model: llama3.1:8b
                choices:
                  - index: 0
                    message:
                      role: assistant
                      content: x402 turns HTTP 402 into a working payment handshake.
                    finish_reason: stop
                usage:
                  prompt_tokens: 24
                  completion_tokens: 12
                  total_tokens: 36
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE frames. Each `data:` line is a JSON
                  `ChatCompletionChunk`; the stream is terminated by the literal frame
                  `data: [DONE]`. Individual frames are validated against
                  `#/components/schemas/ChatCompletionChunk`.
                contentMediaType: text/event-stream
              x-sse-frame:
                description: |
                  Schema every `data:` frame decodes to, except the terminal `data: [DONE]`
                  sentinel. SSE has no standard OpenAPI representation for per-frame schemas,
                  so this extension carries it.
                schema:
                  $ref: "#/components/schemas/ChatCompletionChunk"
              example: |
                data: {"id":"chatcmpl-0f3a9c1f5b2e4d7a","object":"chat.completion.chunk","created":1785772800,"model":"llama3.1:8b","choices":[{"index":0,"delta":{"role":"assistant","content":""},"finish_reason":null}]}

                data: {"id":"chatcmpl-0f3a9c1f5b2e4d7a","object":"chat.completion.chunk","created":1785772800,"model":"llama3.1:8b","choices":[{"index":0,"delta":{"content":"Coins"},"finish_reason":null}]}

                data: {"id":"chatcmpl-0f3a9c1f5b2e4d7a","object":"chat.completion.chunk","created":1785772800,"model":"llama3.1:8b","choices":[{"index":0,"delta":{},"finish_reason":"stop"}]}

                data: [DONE]
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/NoCapacity"
...
components:
  securitySchemes:
    x402Payment:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: |
        Base64 of the JSON `PaymentPayload` defined by x402 v2. Obtain the requirements to sign
        against from the `402` challenge on the same route. The legacy header name
        `Payment-Signature` is also accepted on input.
  parameters:
    XPaymentHeader:
      name: X-PAYMENT
      in: header
      required: false
      description: |
        Base64-encoded JSON `PaymentPayload`. Omit it to receive the `402` challenge that
        describes what to pay.
      schema:
        type: string
        maxLength: 65536
        contentEncoding: base64
        contentMediaType: application/json
        contentSchema:
          $ref: "#/components/schemas/PaymentPayload"
...
  headers:
    PaymentResponse:
      description: |
        Base64-encoded JSON `SettleResponse` for the inbound leg, including the confirmed
        Algorand transaction id.
      schema:
        type: string
        contentEncoding: base64
        contentMediaType: application/json
        contentSchema:
          $ref: "#/components/schemas/SettleResponse"
    MeshNodeId:
      description: Identifier of the node that served this request.
      schema:
        type: string
        maxLength: 128
    MeshRequestId:
      description: Correlation id shared by logs, traces and the settlement record.
      schema:
        type: string
        maxLength: 128
    MeshRouteReason:
      description: Human-readable explanation of why the router picked this node.
      schema:
        type: string
        maxLength: 512
  responses:
    PaymentRequired:
      description: |
        No usable `X-PAYMENT` header was supplied, or the payment failed verification. The body
        states exactly what to pay, to whom, on which chain, and in which asset.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PaymentRequired"
          examples:
            missingPayment:
              summary: First call — no payment attached (Algorand MainNet, $0.0020 USDC)
              value:
                x402Version: 2
                error: X-PAYMENT header is required
                resource:
                  url: https://mesh.example.com/v1/chat/completions
                  description:
                    Pay-per-prompt OpenAI-compatible LLM inference routed to community GPU
                    nodes.
                  mimeType: application/json
                  serviceName: x402 Mesh Inference
                  tags:
                    - ai
                    - inference
                    - llm
                    - algorand
                    - x402-global-challenge
                accepts:
                  - scheme: exact
                    network: algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73k
                    asset: "31566704"
                    amount: "2000"
                    payTo: MESHA65YISXOIW6STKMGJ5AFBZK2UJQNHR3LQZPG2LHIRB4YBPFJPT5WGU
                    maxTimeoutSeconds: 120
                    extra:
                      name: USDC
                      decimals: 6
                      feePayer: ZMFK2OI7ZBD2U27ISERZC4S6LKM6WMFJPZQ4MYNJDZ2VNBNMBA67RA22AA
                extensions:
                  bazaar:
                    info:
                      input:
                        type: http
                        method: POST
                        bodyType: json
                        body:
                          model: llama3.1:8b
                          messages:
                            - role: user
                              content: Explain the x402 protocol in one sentence.
                      output:
                        type: application/json
                        example:
                          id: chatcmpl-0f3a9c1f5b2e4d7a
                          object: chat.completion
                          model: llama3.1:8b
            rejectedPayment:
              summary: Payment supplied but rejected by the facilitator
              value:
                x402Version: 2
                error: "payment rejected: insufficient_funds"
                resource:
                  url: https://mesh.example.com/v1/chat/completions
                  serviceName: x402 Mesh Inference
                accepts:
                  - scheme: exact
                    network: algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73k
                    asset: "31566704"
                    amount: "2000"
                    payTo: MESHA65YISXOIW6STKMGJ5AFBZK2UJQNHR3LQZPG2LHIRB4YBPFJPT5WGU
                    maxTimeoutSeconds: 120
                    extra:
                      name: USDC
                      decimals: 6
                      feePayer: ZMFK2OI7ZBD2U27ISERZC4S6LKM6WMFJPZQ4MYNJDZ2VNBNMBA67RA22AA
    BadRequest:
      description: The request failed schema or freshness validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: validation_error
              message: invalid chat completion request
              details:
                fieldErrors:
                  messages:
                    - messages must not be empty
...
    TooManyRequests:
      description: The caller exceeded its request budget.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: rate_limited
              message: too many requests
    NoCapacity:
      description: No healthy node can serve the requested model right now.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: no_capacity
              message: no healthy node advertises model llama3.1:405b
    BadGateway:
      description: |
        A downstream dependency failed: the selected node returned an unusable response, or an
        on-chain settlement leg could not be completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          examples:
            upstream:
              value:
                error:
                  code: upstream_error
                  message: node returned a malformed completion
            settlement:
              value:
                error:
                  code: settlement_error
                  message: operator payout failed
                  details:
                    reason: ErrConfirmationFailed
    InternalError:
      description: Unexpected server-side failure. The message is deliberately opaque.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MeshError"
          example:
            error:
              code: internal_error
              message: Internal server error
  schemas:
    MeshError:
      type: object
      description: |
        Uniform error envelope. `code` is a stable machine-readable contract; `details` is
        sanitized before serialization, so it never contains key material or a stack trace.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - config_error
                - pricing_error
                - validation_error
                - payment_error
                - settlement_error
                - no_capacity
                - upstream_error
                - auth_error
                - rate_limited
                - internal_error
            message:
              type: string
              maxLength: 1024
            details:
              type: object
              additionalProperties: true
    Network:
      type: string
      description: |
        CAIP-2 network identifier. Both Algorand encodings are accepted on input; responses
        always carry the canonical truncated form.
      minLength: 3
      maxLength: 128
      pattern: "^[a-z0-9-]+:[A-Za-z0-9+/=_-]+$"
      examples:
        - algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73k
        - algorand:SGO1GKSzyE7IEPItTxCByw9x8FmnrCDe
        - algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8=
    AlgorandAddress:
      type: string
      description: 58-character RFC 4648 base32 Algorand address.
      pattern: "^[A-Z2-7]{58}$"
      examples:
        - MESHA65YISXOIW6STKMGJ5AFBZK2UJQNHR3LQZPG2LHIRB4YBPFJPT5WGU
    AtomicAmount:
      type: string
      description: |
        Non-negative integer USDC amount in atomic units (6 decimals) as a decimal string.
        `"2000"` is $0.0020. Never a float.
      pattern: "^\\d+$"
      examples:
        - "2000"
        - "1700"
        - "300"
...
    AssetId:
      type: string
      description: |
        Algorand ASA id as a decimal string. USDC is `31566704` on MainNet and `10458941` on
        TestNet.
      pattern: "^\\d+$"
      examples:
        - "31566704"
        - "10458941"
    ResourceInfo:
      type: object
      description: Descriptive metadata about the paid resource, echoed in the 402 challenge.
      required: [url]
      additionalProperties: false
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
        description:
          type: string
          maxLength: 1024
        mimeType:
          type: string
          maxLength: 128
        serviceName:
          type: string
          maxLength: 128
        tags:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 64
        iconUrl:
          type: string
          format: uri
          maxLength: 2048
    PaymentRequirements:
      type: object
      description: One acceptable way to pay for the resource.
      required: [scheme, network, asset, amount, payTo, maxTimeoutSeconds, extra]
      additionalProperties: false
      properties:
        scheme:
          type: string
          description: Payment scheme. This gateway only offers `exact`.
          examples: [exact]
        network:
          $ref: "#/components/schemas/Network"
        asset:
          $ref: "#/components/schemas/AssetId"
        amount:
          $ref: "#/components/schemas/AtomicAmount"
        payTo:
          $ref: "#/components/schemas/AlgorandAddress"
        maxTimeoutSeconds:
          type: integer
          description: How long the challenge stays valid for.
          minimum: 1
          maximum: 3600
          examples: [120]
        extra:
          type: object
          description: |
            Scheme- and chain-specific metadata. On Algorand this carries the asset `name` and
            `decimals`, plus the facilitator's `feePayer` account, which is what makes the
            payment gasless for the client.
          additionalProperties: true
          properties:
            name:
              type: string
            decimals:
              type: integer
            feePayer:
              $ref: "#/components/schemas/AlgorandAddress"
    PaymentRequired:
      type: object
      description: x402 v2 `402` challenge body.
      required: [x402Version, resource, accepts]
      additionalProperties: false
      properties:
        x402Version:
          type: integer
          const: 2
        error:
          type: string
          description: Why the request was not served.
          maxLength: 1024
        resource:
          $ref: "#/components/schemas/ResourceInfo"
        accepts:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PaymentRequirements"
        extensions:
          type: object
          description: Protocol extensions; `bazaar` carries the discovery declaration.
          additionalProperties: true
    ExactAvmPayload:
      type: object
      description: |
        The Algorand-specific body of a `PaymentPayload` for the `exact` scheme. The group is
        atomic: the facilitator's fee-payer transaction and the client's ASA transfer either
        both commit or both fail.
      required: [paymentGroup, paymentIndex]
      additionalProperties: false
      properties:
        paymentGroup:
          type: array
          description: Base64 msgpack-encoded transactions forming one atomic group.
          minItems: 1
          maxItems: 16
          items:
            type: string
            contentEncoding: base64
        paymentIndex:
          type: integer
          description: Zero-based index of the ASA transfer leg within `paymentGroup`.
          minimum: 0
    PaymentPayload:
      type: object
      description: |
        Decoded contents of the `X-PAYMENT` header. `accepted` must be byte-identical to one of
        the `accepts[]` entries from the challenge.
      required: [x402Version, accepted, payload]
      additionalProperties: false
      properties:
        x402Version:
          type: integer
          const: 2
        resource:
          $ref: "#/components/schemas/ResourceInfo"
        accepted:
          $ref: "#/components/schemas/PaymentRequirements"
        payload:
          $ref: "#/components/schemas/ExactAvmPayload"
        extensions:
          type: object
          additionalProperties: true
    SettleResponse:
      type: object
      description: Facilitator settlement outcome, as carried base64 in `PAYMENT-RESPONSE`.
      required: [success, transaction, network]
      additionalProperties: true
      properties:
        success:
          type: boolean
        errorReason:
          type: string
        errorMessage:
          type: string
        payer:
          $ref: "#/components/schemas/AlgorandAddress"
        transaction:
          type: string
          description: Confirmed Algorand transaction id.
          maxLength: 128
        network:
          $ref: "#/components/schemas/Network"
        amount:
          $ref: "#/components/schemas/AtomicAmount"
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
    ChatCompletionRequest:
      type: object
      description: |
        OpenAI-compatible request. Unknown properties are ignored rather than rejected, so
        drop-in OpenAI SDK clients work unchanged.
      required: [model, messages]
      additionalProperties: true
      properties:
        model:
          type: string
          minLength: 1
          maxLength: 200
        messages:
          type: array
          minItems: 1
          maxItems: 512
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
          description: When true the response is `text/event-stream`.
          default: false
        max_tokens:
          type: integer
          minimum: 1
          maximum: 1000000
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        stop:
          type: array
          maxItems: 4
          items:
            type: string
    ChatCompletionUsage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
          minimum: 0
        completion_tokens:
          type: integer
          minimum: 0
        total_tokens:
          type: integer
          minimum: 0
    ChatCompletionChoice:
      type: object
      required: [index, message, finish_reason]
      properties:
        index:
          type: integer
          minimum: 0
        message:
          $ref: "#/components/schemas/ChatMessage"
        finish_reason:
          type: [string, "null"]
          examples: [stop, length]
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          minLength: 1
        object:
          type: string
          const: chat.completion
        created:
          type: integer
          description: Unix epoch seconds.
          minimum: 0
        model:
          type: string
          minLength: 1
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/ChatCompletionUsage"
    ChatCompletionChunkChoice:
      type: object
      required: [index, delta, finish_reason]
      properties:
        index:
          type: integer
          minimum: 0
        delta:
          type: object
          description: Partial `ChatMessage`; either field may be absent on any given frame.
          properties:
            role:
              type: string
              enum: [system, user, assistant]
            content:
              type: string
        finish_reason:
          type: [string, "null"]
    ChatCompletionChunk:
      type: object
      description: One SSE frame of a streamed completion.
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
          minLength: 1
        object:
          type: string
          const: chat.completion.chunk
        created:
          type: integer
          minimum: 0
        model:
          type: string
          minLength: 1
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChunkChoice"
...